requirements.txt
discord.py
openai
python-dotenv
orjson
aiosqlite
